refactor crypto_stream tests
